Split apart action data for code organization
